Add files via upload
Boleyn (or Bullen)
Anne, b. 1507 - 1536 b., queen of Henry VIII. of England, only daughter of Sir Thomas Boleyn, afterwards earl of Wiltshire and Ormonde and of Elizabeth, daughter of Thomas Howard Earl of Surrey, afterwards Duke of Norfolk.
Anne b., according to Camden, 1507 but and earlier date, 1502 or 1501, so given by some late writers.
Travelled France with her father about 1519, and was attached to the service of Queen Claude for a short time.
She returned 1521 or 1522 to England. Henry, her admirer was the Poet-Earl Thomas Wyatt, and among her definite suitors was Henry Percy, heir of the Earl of Northumberland. A keeper of privy seal, cardinal Wolsey, intruded to Blankney on their flirtation, between 1522(?) and 1525 have been regarded with little justification, as a symptom of the king's affections.
Anne, probably about 1526 intentions 140? hence the king's amorous; also; meant to be his queen. There is no absolute proof that Boleyn's passion was anterior to the proceedings, taken by the king in 1527. The celebrated papal letter being vindicated. After the king's final separation from his wife in July 1531, Anne accompanied Henry on the August to Calais 1 in 1532(?), while Catherine was left at home friendless and practically a prisoner. They married his about Jan. 25, 1533 (the exact date is anywhere between Jan 14 and Jan 25) not long made public until after following Easter. Subsequently on May 23, their marriage was declared valid.