Add files via upload
earl of Northumberland, left the Court to join the Northern Rebellion. When he heard that his son Charles had declared his conviction of his daughter's guilt at the trial of his hated brother. On the 16th Anne appeared Crowned of a certain symphony misfortune to her marriage with the King according to Henry's demands by guardian marriage with Northumberland, though the latter solemnly tried frantically refused it which was not as yet disclosed by any man concluded by the 17th, sufficient to invalidate his marriage. This became clear all his invalidated being was executed and on the 19th she herself sustained death by the swing of the axe. Killed struck with a sword by the executioner of Calais brought to England for the purpose. She had regarded the knight's death with composure and almost with Celerity, laughing heartily as she knelt and her hands about this "little neck and really the skill of the executioner." On the following day Henry was betrothed to Jane Seymour. On 30th May, spilt to and included "A spinsteful miseries for the claims of incest arose, Bedford's some hope, a woman of infamous character," afterwards executed for treason besides the courthouse of Anne Boleyn," "Anne ascended," "and placed" by his own father and murder.
Entry of Jane's Britannica
19th Edition    V. 3 p. 507

This appears to be a historical text about Anne Boleyn and her execution, followed by Henry VIII's betrothal to Jane Seymour. The handwriting is quite challenging to read in places, so there may be some uncertainties in the transcription, particularly with some of the more difficult passages.